# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RateMaster API",
    "description": "Revenue/pricing recommendation SaaS for hotels",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/jobs/engine-a": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Trigger Engine A",
        "description": "Trigger Engine A (tactical 0-30 days) run.",
        "operationId": "trigger_engine_a_api_v1_jobs_engine_a_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerEngineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/engine-b": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Trigger Engine B",
        "description": "Trigger Engine B (strategic 31-365 days) run.",
        "operationId": "trigger_engine_b_api_v1_jobs_engine_b_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerEngineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/status/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Status",
        "description": "Get job status - poll for progress.",
        "operationId": "job_status_api_v1_jobs_status__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/training": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Trigger Training",
        "description": "Trigger training job for property - builds dataset, trains, registers model.",
        "operationId": "trigger_training_api_v1_jobs_training_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerEngineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/engines/runs": {
      "get": {
        "tags": [
          "engines"
        ],
        "summary": "List Engine Runs",
        "description": "List engine runs for user's properties (owner, GM, or Analyst).",
        "operationId": "list_engine_runs_api_v1_engines_runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Property Id"
            }
          },
          {
            "name": "engine_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Engine Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngineRunResponse"
                  },
                  "title": "Response List Engine Runs Api V1 Engines Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/engines/runs/{run_id}": {
      "get": {
        "tags": [
          "engines"
        ],
        "summary": "Get Engine Run",
        "description": "Get engine run with recommendations.",
        "operationId": "get_engine_run_api_v1_engines_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineRunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/engines/recommendations/{rec_id}/apply": {
      "post": {
        "tags": [
          "engines"
        ],
        "summary": "Mark Recommendation Applied",
        "description": "Mark recommendation as applied. Owner and GM can approve; Analyst cannot.",
        "operationId": "mark_recommendation_applied_api_v1_engines_recommendations__rec_id__apply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rec_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rec Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EngineRunDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "property_id": {
            "type": "string",
            "title": "Property Id"
          },
          "engine_type": {
            "type": "string",
            "title": "Engine Type"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "recommendations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendations"
          },
          "calendar": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "engine_type",
          "run_id",
          "status",
          "confidence",
          "created_at"
        ],
        "title": "EngineRunDetailResponse"
      },
      "EngineRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "property_id": {
            "type": "string",
            "title": "Property Id"
          },
          "engine_type": {
            "type": "string",
            "title": "Engine Type"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "engine_type",
          "run_id",
          "status",
          "confidence",
          "created_at"
        ],
        "title": "EngineRunResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "stay_date": {
            "type": "string",
            "title": "Stay Date"
          },
          "suggested_bar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Bar"
          },
          "current_bar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Bar"
          },
          "delta_dollars": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Dollars"
          },
          "delta_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Pct"
          },
          "occupancy_projection": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupancy Projection"
          },
          "occupancy_projection_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupancy Projection Low"
          },
          "occupancy_projection_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupancy Projection High"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "why_bullets": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Bullets"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stay_date",
          "suggested_bar",
          "current_bar",
          "delta_dollars",
          "delta_pct",
          "occupancy_projection",
          "confidence",
          "why_bullets",
          "applied"
        ],
        "title": "RecommendationResponse"
      },
...
      "TriggerEngineRequest": {
        "properties": {
          "property_id": {
            "type": "string",
            "title": "Property Id"
          }
        },
        "type": "object",
        "required": [
          "property_id"
        ],
        "title": "TriggerEngineRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}